requirements.txt
annotated-types==0.7.0
anyio==4.5.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fsspec==2024.9.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
nltk==3.9.1
numpy==1.24.4
openai==1.51.0
packaging==24.1
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.1.1
sentencepiece==0.2.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.0
torch==2.2.2
tqdm==4.66.5
transformers==4.45.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
